Fake_News_Detection/requirements.txt

streamlit
pytesseract
Pillow
scikit-learn
numpy
